Move wiki specific match input processing to dedicated modules
